refactor : feature 폴더 내부에 query 폴더 생성 (#116)
