+++ src/features/clinic/query/queryKeys.ts
@@ -1,0 +1,6 @@
+export const queryKeys = {
+  all: ['clinic'] as const,
+  detail: (medicalId: number) => [...queryKeys.all, 'detail', medicalId] as const,
+  calendar: (userId: string, localDate: string) =>
+    [...queryKeys.all, 'calendar', userId, localDate] as const,
+}
