+++ src/features/clinic/query/useClinicApi.ts
@@ -6,15 +6,9 @@ import {
   createClinicInfo,
   deleteClinicInfo,
   editClinicInfo,
-} from '@/apis'
+  queryKeys,
+} from '@/features'
 import type { ClinicCalendarRequest } from '@/types'
-
-const queryKeys = {
-  all: ['clinic'] as const,
-  detail: (medicalId: number) => [...queryKeys.all, 'detail', medicalId] as const,
-  calendar: (userId: string, localDate: string) =>
-    [...queryKeys.all, 'calendar', userId, localDate] as const,
-}
 
 export const useClinicInfo = (medicalId: number) =>
   useQuery({
