+++ src/features/clinic/service/useInsertedClinicForm.ts
@@ -4,7 +4,7 @@ import { useParams } from 'next/navigation'
 import dayjs from 'dayjs'
 
 import { ALARM_TIME } from '@/constants'
-import { useClinicInfo } from '@/hooks'
+import { useClinicInfo } from '@/features'
 import { useClinicForm } from '@/schema'
 
 /**
