+++ src/features/index.ts
@@ -11,6 +11,9 @@ export * from './sign-up/UserInfoStep'
 export * from './clinic/hook/useFormattedVisitDate'
 export * from './clinic/hook/useSelectAlarmTime'
 export * from './clinic/hook/useTagToggle'
+export * from './clinic/query/clinicApi'
+export * from './clinic/query/queryKeys'
+export * from './clinic/query/useClinicApi'
 export * from './clinic/service/useInsertedClinicForm'
 export * from './clinic/service/useMonthlyScheduleList'
 export * from './clinic/service/useSubmitAddClinicForm'
